video: move video handling to the video/ folder
