+++ src/font/font_freetype.c
@@ -34,7 +34,7 @@
 #include "font.h"
 #include "shl_log.h"
 #include "shl_misc.h"
-#include "uterm_video.h"
+#include "video/video.h"
 
 #define LOG_SUBSYSTEM "font_freetype"
