+++ src/font/font_unifont.c
@@ -47,7 +47,7 @@
 #include "font.h"
 #include "font_unifont_data.bin.h"
 #include "shl_log.h"
-#include "uterm_video.h"
+#include "video/video.h"
 
 #define LOG_SUBSYSTEM "font_unifont"
