+++ src/kmscon_dummy.c
@@ -34,7 +34,7 @@
 #include "kmscon_seat.h"
 #include "shl_dlist.h"
 #include "shl_log.h"
-#include "uterm_video.h"
+#include "video/video.h"
 
 #define LOG_SUBSYSTEM "dummy"
