+++ src/kmscon_seat.c
@@ -45,8 +45,8 @@
 #include "kmscon_terminal.h"
 #include "shl_dlist.h"
 #include "shl_log.h"
-#include "uterm_video.h"
 #include "uterm_vt.h"
+#include "video/video.h"
 
 #define LOG_SUBSYSTEM "seat"
