+++ src/text.c
@@ -39,7 +39,7 @@
 #include "shl_misc.h"
 #include "shl_register.h"
 #include "text.h"
-#include "uterm_video.h"
+#include "video/video.h"
 
 #define LOG_SUBSYSTEM "text"
