+++ src/video/drm2d_render.c
@@ -37,12 +37,12 @@
 #include <unistd.h>
 #include <xf86drm.h>
 #include <xf86drmMode.h>
+#include "drm2d_internal.h"
+#include "drm_shared_internal.h"
 #include "eloop.h"
 #include "shl_log.h"
-#include "uterm_drm2d_internal.h"
-#include "uterm_drm_shared_internal.h"
-#include "uterm_video.h"
-#include "uterm_video_internal.h"
+#include "video.h"
+#include "video_internal.h"
 
 #define LOG_SUBSYSTEM "uterm_drm2d_render"
