+++ src/video/drm3d_video.c
@@ -45,15 +45,15 @@
 #include <unistd.h>
 #include <xf86drm.h>
 #include <xf86drmMode.h>
+#include "drm3d_internal.h"
+#include "drm_shared_internal.h"
 #include "eloop.h"
 #include "shl_log.h"
 #include "shl_misc.h"
-#include "uterm_drm3d_internal.h"
-#include "uterm_drm_shared_internal.h"
-#include "uterm_video.h"
-#include "uterm_video_internal.h"
+#include "video.h"
+#include "video_internal.h"
 
-#define LOG_SUBSYSTEM "uterm_drm3d_video"
+#define LOG_SUBSYSTEM "drm3d_video"
 
 static void bo_destroy_event(struct gbm_bo *bo, void *data)
 {
